+++ coterie-app/ai_models/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class AiModelsConfig(AppConfig):
+    name = 'ai_models'
