+++ coterie-app/ai_models/messages.py
@@ -1,0 +1,5 @@
+MODEL_EXISTS_ERROR = "a model with name {} already exists!"
+MODEL_NOT_FOUND = "No Model with is id was found."
+FIELD_KEY_ERROR = "{} field must be provided!"
+HTTP_BAD_REQUEST = {"Error": "Bad Request!"}
+HTTP_NOT_FOUND = {"Error": "Not Found!"}
