+++ coterie-app/ai_models/tests.py
