--- coterie-app/api/serializers.py
@@ -1,153 +1,0 @@
-from rest_framework import serializers
-from django.conf import settings
-import json
-import requests
-from .models import Team, Enrollments, Project, AiModel, Deployment
-from .utils import Roles, default_role, create_job
-from authentication.models import User
-
-
-class TeamMemberSerializer(serializers.ModelSerializer):
-    #team = serializers.CharField(source='team.name')
-    #user = serializers.CharField(source='user.id')
-    email = serializers.CharField(source='user.email')
-
-    class Meta:
-        model = Enrollments
-        fields = ['user', 'email', 'role', 'status']
-
-
-# class TeamUserSerializer(serializers.ModelSerializer):
-#     class Meta:
-#         model = User
-#         fields = ['']
-
-class TeamSerializer(serializers.ModelSerializer):
-    members = serializers.SerializerMethodField()
-
-    def get_members(self, obj):
-        members = Enrollments.objects.filter(team=obj.id)
-        serializer = TeamMemberSerializer(members, many=True)
-        return serializer.data
-
-    class Meta:
-        model = Team
-        fields = ['id', 'name', 'owner', 'created_at', 'members', 'updated_at']
-        extra_kwargs = {'name': {'required': True}, 'owner': {
-            'read_only': True}, 'members': {'required': False}}
-
-
-class TeamInviteSerializer(serializers.ModelSerializer):
-    role = serializers.ChoiceField(choices=Roles, default=default_role)
-    email = serializers.ReadOnlyField(source='user.email')  # write_only=True
-    invite_email = serializers.EmailField(write_only=True)
-
-    class Meta:
-        model = Enrollments
-        fields = ['id', 'email', 'invite_email', 'user', 'team', 'role',
-                  'status', 'created_at', 'updated_at']
-        extra_kwargs = {'status': {'read_only': True},
-                        'user': {'read_only': True}}
-
-    def validate(self, attrs):
-        #team = Team.objects.get(id=attrs['team'].id)
-        # if not team.owner == self.context.get('request').user:
-        #    raise serializers.ValidationError({"error":"only team creators can make invites"})
-
-        if not User.objects.filter(email=attrs['invite_email']).exists():
-            raise serializers.ValidationError(
-                {"error": "no user matching this email was found"})
-        return attrs
-
-    def create(self, validated_data):
-        # eliminate the email filed as it is not in the model.
-        validated_data.pop('invite_email', None)
-
-        return super().create(validated_data)
-
-
-class TeamMemberUpdateSerializer(serializers.ModelSerializer):
-    role = serializers.ChoiceField(choices=Roles, default=default_role)
-
-    class Meta:
-        model = Enrollments
-        fields = ['id', 'user', 'team', 'role',
-                  'status', 'created_at', 'updated_at']
-        extra_kwargs = {'status': {'read_only': True}}
-
-    # def validate(self, attrs):
-    #    team = Team.objects.get(id=attrs['team'].id)
-    #    if not team.owner == self.context.get('request').user:
-    #        raise serializers.ValidationError({"error":"only team creators can make invites"})
-    #
-    #    return attrs
-
-
-class EmailVerificationSerializer(serializers.Serializer):
-    token = serializers.CharField(max_length=555)
-
-    class Meta:
-        fields = ['token']
-
-# projects should have two urls, 1 serializer, two views
-# we should dynamically identify the team and the creator of a project
-# only team creator or users with owner role can create/delete projects
-# only invited peeps have priviledge of viewing list and detail views
-# contributors can only update, they can neither create nor destroy projects
-# List/Create -> api/team/projects
-# retrieve, update, destroy -> api/team/project/id
-
-
-class ProjectSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Project
-        fields = ['id', 'title', 'description', "is_archived",
-                  'url', 'team', 'creator', 'created_at', 'updated_at']
-        extra_kwargs = {'creator': {'read_only': True}, 'url': {'required': False}, 'team': {
-            'read_only': True}, 'description': {'required': False}, "is_archived": {"required": False, "default": False}}
-
-
-class AiModelSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = AiModel
-        fields = ['id', 'model_name', 'gcr_url', 'created_at', 'updated_at']
-        read_only_fields = ['created_at', 'updated_at']
-
-
-class SeldonDeploymentSerializer(serializers.Serializer):
-    name = serializers.CharField(max_length=555, read_only=True)
-
-    class Meta:
-        name = ['name']
-
-
-class DeploymentSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Deployment
-        fields = ['id', 'deployment_id', 'name', 'model', 'project',
-                  'description', 'service_endpoint', 'creator', 'created_at', 'updated_at']
-        read_only_fields = ['deployment_id', 'project', 'creator',
-                            'service_endpoint', 'created_at', 'updated_at']
-
-    def create(self, validated_data):
-        # get the json created
-        name = validated_data.get('name')
-        json_job = create_job(name)
-        # get the image url updated
-        json_job['spec']['predictors'][0]['componentSpecs'][0]['spec']['containers'][0]['image'] = validated_data.get(
-            'model').gcr_url
-        # make a post to producer api
-        json_job = json.dumps(json_job)
-        url = settings.KAFKA_API_URL
-        res = requests.post(url, data=json_job)
-        try:
-            validated_data['deployment_id'] = res.json()['model_name']
-            return super().create(validated_data)
-        except:
-            raise serializers.ValidationError(
-                {"error": "unable to reach producer api successfully"})
-
-    def update(self, instance, validated_data):
-        instance.description = validated_data.get(
-            "description", instance.description)
-        return instance
