+++ coterie-app/app/settings.py
@@ -47,7 +47,8 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     'drf_yasg',
     'corsheaders',
-    'api'
+    'ai_models',
+    'teams_api',
 ]
 
 MIDDLEWARE = [
