+++ coterie-app/teams_api/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TeamsApiConfig(AppConfig):
+    name = 'teams_api'
