+++ coterie-app/teams_api/custompermissions.py
@@ -1,9 +1,6 @@
 from rest_framework.permissions import BasePermission, SAFE_METHODS
-from .models import Project, Team, Enrollments
+from .models import Team, Enrollments
 from .utils import Status as user_status, Roles
-# -> only team members can view a project
-# -> only project creator and team owner can delete project
-# -> when you're in a team and ur role is not viewer, u can update a project
 
 
 class hasTeamDetailPermissions(BasePermission):
@@ -40,6 +37,35 @@ class hasTeamDetailPermissions(BasePermission):
         if request.method in ['DELETE']:
 
             # -> no admin priviledge here: request.user.is_superuser
+            return is_owner or has_owner_priviledge
+
+
+class CanInviteUser(BasePermission):
+    """only owner and user with owner privildege can ivite or delete an invited member"""
+    message = "only team members and owners have this permission"
+
+    def has_permission(self, request, view):
+        is_owner = Team.objects.filter(
+            owner=request.user.id, id=view.kwargs['team_id']).exists()
+        is_member = Team.objects.filter(
+            id=view.kwargs['team_id'], members__id=request.user.id).exists()
+
+        if is_member:
+            self.message = "please accept team invite first."
+            team_status_obj = Enrollments.objects.get(
+                team=view.kwargs['team_id'], user=request.user.id)
+            is_verified_member = team_status_obj.status == user_status.ACCEPTED
+            if is_verified_member:
+                has_owner_priviledge = team_status_obj.role == Roles.OWNER
+        else:
+            is_verified_member = False
+            has_owner_priviledge = False
+
+        if request.method in SAFE_METHODS:
+            return is_owner or is_verified_member
+
+        else:
+            self.message = "only user with owner priviledge can perform this action"
             return is_owner or has_owner_priviledge
 
 
@@ -86,42 +112,3 @@ class IsOwnerOrContributor(BasePermission):
            # DELETE -> owner , creator
             self.message = "only user with owner priviledge can perform this action"
             return is_owner or has_owner_priviledge
-
-
-class CanInviteUser(BasePermission):
-    """only owner and user with owner privildege can ivite or delete an invited member"""
-    message = "only team members and owners have this permission"
-
-    def has_permission(self, request, view):
-        is_owner = Team.objects.filter(
-            owner=request.user.id, id=view.kwargs['team_id']).exists()
-        is_member = Team.objects.filter(
-            id=view.kwargs['team_id'], members__id=request.user.id).exists()
-
-        if is_member:
-            self.message = "please accept team invite first."
-            team_status_obj = Enrollments.objects.get(
-                team=view.kwargs['team_id'], user=request.user.id)
-            is_verified_member = team_status_obj.status == user_status.ACCEPTED
-            if is_verified_member:
-                has_owner_priviledge = team_status_obj.role == Roles.OWNER
-        else:
-            is_verified_member = False
-            has_owner_priviledge = False
-
-        if request.method in SAFE_METHODS:
-            return is_owner or is_verified_member
-
-        else:
-            self.message = "only user with owner priviledge can perform this action"
-            return is_owner or has_owner_priviledge
-
-
-class IsAdminUserOrReadonly(BasePermission):
-    message = "only admins can create or modify models"
-
-    def has_permission(self, request, view):
-        is_admin = request.user.is_superuser
-        if request.method in SAFE_METHODS:
-            return True
-        return is_admin
